requirements.txt
pandas
python-dotenv
mysql-connector-python
pymongo
